Update TriviaQuestionUI.cs
Changed the Dynamic of the Infinite Loop in order to allow the File to Load. In the case of this not working, there might be a need to multi-thread this idea, and that is hard to do, if not unnecessary, because dealing with the thread is hazards. C# is no the language to do threading in.